sbus: store dbus connection name in domain.conn_name
We're moving towards a single D-Bus server in monitor, therefore we
will be relying on connection names instead of connection addresses.

It makes sense to have the name easily available from the domain.
