+++ src/confdb/confdb.c
@@ -22,6 +22,7 @@
 #include "config.h"
 
 #include <ctype.h>
+#include "sbus/sbus_opath.h"
 #include "util/util.h"
 #include "confdb/confdb.h"
 #include "confdb/confdb_private.h"
@@ -906,6 +907,29 @@ static char *confdb_get_domain_hostname(TALLOC_CTX *mem_ctx,
     return talloc_strdup(mem_ctx, sys);
 }
 
+static char *confdb_get_domain_bus(TALLOC_CTX *mem_ctx,
+                                   struct sss_domain_info *domain)
+{
+    struct sss_domain_info *head;
+    char *safe_name;
+    char *bus_name;
+
+    /* There is only one bus that belongs to the top level domain. */
+    head = get_domains_head(domain);
+
+    safe_name = sbus_opath_escape(mem_ctx, head->name);
+    if (safe_name == NULL) {
+        return NULL;
+    }
+
+    /* Parts of bus names must not start with digit thus we concatenate
+     * the name with underscore instead of period. */
+    bus_name = talloc_asprintf(mem_ctx, "sssd.domain_%s", safe_name);
+    talloc_free(safe_name);
+
+    return bus_name;
+}
+
 static errno_t confdb_init_domain(struct sss_domain_info *domain,
                                   struct ldb_result *res)
 {
@@ -924,7 +948,6 @@ static errno_t confdb_init_domain(struct sss_domain_info *domain,
         ret = ENOMEM;
         goto done;
     }
-    domain->conn_name = domain->name;
 
     tmp = ldb_msg_find_attr_as_string(res->msgs[0],
                                       CONFDB_DOMAIN_ID_PROVIDER,
@@ -1047,6 +1070,12 @@ static errno_t confdb_init_domain(struct sss_domain_info *domain,
         }
     }
 #endif
+
+    domain->conn_name = confdb_get_domain_bus(domain, domain);
+    if (!domain->conn_name) {
+        ret = ENOMEM;
+        goto done;
+    }
 
     domain->not_found_counter = 0;
     ret = EOK;
